Set up `series/0.2` branch (#215)
This is a preparation for a 'courtesy' release based on http4s 0.22.14 (and Cats-Effect 2). I've created a new branch `series/0.2` (it shouldn't be merged to `main`!) just like we do in http4s. Why this is needed? Well, because many users are still on Cats-Effect 2 and can't invest in migration to Cats-Effect 3, but wants to get new functionality of http4s. I believe this is a one-time story, we'll publish `0.2.1` and can remove the `series/0.2` branch.